requirements.txt
# Forecast - Dependencies
# =======================

# Core Dash
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-ag-grid>=31.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0

# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0

# Server
gunicorn>=21.0.0
